refactor: extracted inline component
