+++ src/views/components/SearchNavbar/ISearchNavbar.tsx
@@ -1,0 +1,8 @@
+import { Genre, Publisher, Tag } from '@data/types'
+
+
+export interface ISearchNavbar {
+	genres: Genre[]
+	tags: Tag[]
+	publishers: Publisher[]
+}
