+++ src/views/components/SearchWidget/SearchWidget.tsx
@@ -1,10 +1,15 @@
 import { Genre, Publisher, Tag } from '@data/types'
 import { GenreService, PublisherService, TagService } from '@src/services'
 import { useEffect, useState } from 'react'
-import { Link } from 'react-router-dom'
 import style from './SearchWidget.module.scss'
 import { SearchForm } from '@views/components/SearchForm/SearchForm.tsx'
+import {
+	SearchNavbar
+} from '@views/components/SearchNavbar/SearchNavbar.tsx'
+import { ISearchNavbar } from '@views/components/SearchNavbar/ISearchNavbar.tsx'
 
+
+export { SearchWidget }
 
 function SearchWidget() {
 	const [genres, setGenres]         = useState<Genre[]>([])
@@ -32,39 +37,3 @@ function SearchWidget() {
 	)
 }
 
-interface ISearchNavbar {
-	genres: Genre[]
-	tags: Tag[]
-	publishers: Publisher[]
-}
-
-function SearchNavbar(props: ISearchNavbar) {
-	const getList = (data: Genre[] | Tag[] | Publisher[]) => data.map(d => {
-		const key = `search-${typeof (d)}-${d.id}`
-
-		return (<li key={key}><Link to="/search">{d.name}</Link></li>)
-	})
-
-	const getSection = (key: keyof ISearchNavbar) => {
-		return (
-			<li>
-				<details>
-					<summary>{key}</summary>
-
-					<ul className={style.SearchOptions}>
-						{getList(props[key])}
-					</ul>
-				</details>
-			</li>)
-	}
-
-	return (
-		<menu className={style.SearchMenu}>
-			{getSection('genres')}
-			{getSection('tags')}
-			{getSection('publishers')}
-		</menu>
-	)
-}
-
-export { SearchWidget }
